Update xlf for clang@14.0.5.ibm. A hack to move forward.
Normally, we build the asan test target with regular clang.  It turns
out that with the new naming of compilers: clang@15.0.0(.ibm,.gcc.8.3.1)
spack is confused and may not use the chosen compiler...

I set xlf to the same value in all four compilers for now, but the issue
is deeper than that and will be reported to Spack developers.