rpath to /opt/cray/pe/mpich/8.1.31/gtl/lib in some tioga compilers
Make sure libmpi_gtl_hsa.so.0 is found when using cray-mpich 8.1.31
by adding rpath to /opt/cray/pe/mpich/8.1.31/gtl/lib to compilers
associated to that version of cray-mpich.

This is specific to cray-mpich 8.1.31, and is due to a missing
symlink with that version, hopefully fixed in subsequent ones.

We noticed the error in Caliper CI jobs.